requirements.txt
absl-py==1.0.0
affine==2.3.1
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
av==8.0.3
backcall==0.2.0
basemap-data==1.3.2
beautifulsoup4==4.11.1
bleach==4.1.0
cachetools==5.0.0
captum==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
clip==1.0
colorama==0.4.6
conda==4.3.16
contourpy==1.0.6
cvbase==0.5.5
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dgl==1.0.0
docker-pycreds==0.4.0
dominate==2.6.0
dtaidistance==2.3.6
earthengine-api==0.1.316
einops==0.4.1
entrypoints==0.3
exceptiongroup==1.0.4
faiss-cpu==1.7.2
fbpca==1.0
ffmpeg-python==0.2.0
Fiona==1.9.1
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5
gast==0.5.3
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-api-python-client==1.12.11
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.43.0
gurobipy==10.0.1
h11==0.14.0
h5py==3.6.0
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
imageio==2.13.0
imageio-ffmpeg==0.4.5
imagesize==1.3.0
imgcat==0.5.0
importlib-metadata==4.10.1
iopath==0.1.9
ipdb==0.13.9
ipykernel==6.7.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lark==1.0.0
libclang==13.0.0
llvmlite==0.38.0
lmdb==1.3.0
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
multidict==6.0.2
munch==2.5.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
novograd==0.0.6
numba==0.55.1
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.4.60
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parallelbar==1.3.1
parso==0.8.3
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.0
portalocker==2.3.2
prettytable==3.4.1
proglog==0.1.9
prometheus-client==0.13.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycosat==0.6.3
pycparser==2.21
pydensecrf==1.0rc2
Pygments==2.10.0
pyparsing==3.0.6
pyproj==3.4.0
pyrsistent==0.18.1
pyshp==2.3.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.1.8
pytorch-lightning-bolts==0.3.0
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
qudida==0.0.4
rasterio==1.2.10
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.6.0
Send2Trash==1.8.0
sentry-sdk==1.6.0
setproctitle==1.2.3
setuptools==58.0.4
setuptools-scm==6.3.2
shapely==2.0.1
shortuuid==1.0.9
SimpleCRF==0.2.1.1
six==1.16.0
sklearn-crfsuite==0.3.6
sktime==0.8.0
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
svgwrite==1.4.3
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.3.2
toml==0.10.2
tomli==1.2.2
torch==1.7.1
torchaudio==0.10.0
torchfile==0.1.0
torchmetrics==0.10.3
torchvision==0.8.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
Tree==0.2.4
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
visdom==0.1.8.9
wandb==0.12.21
wcwidth==0.2.5
webdriver-manager==3.8.4
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yacs==0.1.8
yarl==1.7.2
zipp==3.7.0
